Controllo esistenza file Log

+++ bash_generate_bck_app.sh
@@ -16,10 +16,12 @@
 #==========================================================================================
 
 logger(){
-	SIZE_FILE=$( stat -c %s ${LOG_FILE})
-	if [ ${SIZE_FILE} -gt 1048576 ]; then
-		mv ${LOG_FILE} "${LOG_FILE}.`date +%d%m%Y_%H:%M`"
-	fi
+	if [ -f ${LOGFILE} ]; then
+                SIZE_FILE=$( stat -c %s ${LOG_FILE})
+                if [ ${SIZE_FILE} -gt 1048576 ]; then
+                        mv ${LOG_FILE} "${LOG_FILE}.`date +%d%m%Y_%H:%M`"
+                fi
+        fi
 	CURRENT_TIMESTAMP=`date "+%d/%m/%Y %H:%M:%S"`
 	
 	echo ${CURRENT_TIMESTAMP} [$1] $2 >> ${LOG_FILE}
